+++ app/src/main/kotlin/com/architects/pokearch/ui/details/DetailScreen.kt
@@ -1,28 +1,39 @@
 package com.architects.pokearch.ui.details
 
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import coil.compose.rememberAsyncImagePainter
+import com.architects.pokearch.ui.components.extensions.GetColorsBackground
+import com.architects.pokearch.ui.components.extensions.buildImageRequest
 import com.architects.pokearch.ui.components.image.PokeArchAsyncImage
 import com.architects.pokearch.ui.components.progressIndicators.PokeArchLoadingIndicator
 import com.architects.pokearch.ui.details.state.DetailUiState
-import com.architects.pokearch.ui.home.buildImageRequest
 
 @Composable
 fun DetailScreen(
@@ -57,12 +68,7 @@ fun DetailScreen(
 private fun DetailSuccessScreen(state: DetailUiState.Success) {
 
     Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(
-                horizontal = 16.dp,
-                vertical = 8.dp
-            ),
+        modifier = Modifier.fillMaxSize(),
         propagateMinConstraints = true,
     ) {
 
@@ -70,16 +76,37 @@ private fun DetailSuccessScreen(state: DetailUiState.Success) {
             model = state.pokemonInfo.getImageUrl().buildImageRequest(LocalContext.current)
         )
 
+        val colorDefault = MaterialTheme.colorScheme.surfaceVariant
+        val colorsDefault = listOf(colorDefault, colorDefault)
+
+        var colors by remember { mutableStateOf(colorsDefault) }
+
+        image.GetColorsBackground {
+            colors = it.ifEmpty { colorsDefault }
+        }
+
         Column(
             horizontalAlignment = Alignment.CenterHorizontally,
             verticalArrangement = Arrangement.Top,
         ) {
-            Spacer(modifier = Modifier.size(62.dp))
-            PokeArchAsyncImage(
-                modifier = Modifier.size(260.dp),
-                asyncImagePainter = image,
-                contentDescription = null
-            )
+            Box(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .clip(RoundedCornerShape(
+                        topStart = 0.dp,
+                        topEnd = 0.dp,
+                        bottomStart = 24.dp,
+                        bottomEnd = 24.dp
+                    ))
+                    .background(Brush.linearGradient(colors))
+            ) {
+                PokeArchAsyncImage(
+                    modifier = Modifier.size(240.dp),
+                    asyncImagePainter = image,
+                    contentDescription = null
+                )
+            }
+
             Text(
                 text = state.pokemonInfo.capitalizedName(),
                 style = MaterialTheme.typography.headlineLarge
